User can log out, session is cleared
