+++ app.rb
@@ -4,6 +4,7 @@ require_relative './app/models/user'
 
 class Chitter < Sinatra::Base
 
+  use Rack::MethodOverride
   register Sinatra::Flash
   enable :sessions
   set :session_secret, 'super secret'
@@ -35,6 +36,12 @@ class Chitter < Sinatra::Base
        flash.now[:errors] = ['The email or password is incorrect']
        erb(:'sessions/new')
      end
+   end
+
+   delete '/sessions' do
+     session[:user_id] = nil
+     flash.keep[:notice] = 'Bye!'
+     redirect to('/home')
    end
 
   get '/users/new' do
