+++ spec/user_sign_up_spec.rb
@@ -46,17 +46,4 @@ feature 'I want to sign up for Chitter' do
     expect(page).to have_content("Username is already taken")
   end
 
-  def sign_up(username: "tansaku",
-            name: "Samuel Russell Hampden Joseph",
-            email: "sam@makersacademy.com",
-            password: "s3cr3t",
-            password_confirmation: "s3cr3t")
-    visit "/users/new"
-    fill_in :username, with: username
-    fill_in :name, with: name
-    fill_in :email, with: email
-    fill_in :password, with: password
-    fill_in :password_confirmation, with: password_confirmation
-    click_button "Submit"
-  end
 end
